[feat] implement dynamic routing for the CaseSummaryScreen.
